Swift4 Support.
Due to SE-0110, all of the binder arity functions become ambiguous with arity-1 and arity-0 to the swift compiler. As a solution in order to unblock using Cleanse with Swift4, renaming just these two functions to include their arity count (`to0` and `to1`) minimizes the surface area of the public API that requires modification. This also makes it much easier to revert back to simply using `to` when the issue is resolved in Swift.

Something else worth noting is that you will receive a runtime error (missing binding) if you use the incorrect function name for `to0` or `to1`.