Swift 4 support improvement.
Previous swift 4 fix was to rename arity-0 and arity-1 factory functions to `to0` and `to1` respectively; however, I discovered a better fix that only involves adding extra parenthesis around the factory args for arity-1. This has the great benefit of being able to deprecate the patched function names to work with the swift 4 compiler.