Retain a strong reference to all Future Providers.
If the type was of a `AnyWeakProvider` type, then we would only retain a weak reference to its `FutureProvider` instance causing it to become deallocated. We want to retain a strong reference to this `FutureProvider` like we do for normal providers because we still need to be able to acquire an instance of its getter.