Retain subcomponent graphs.
Subcomponents today are fairly buggy. `Graph` instances that are required to construct subcomponents cause exceptions on `build(:)`. In order to fix this, we need something to retain `Graph` object instances.

This change puts the retention responsibility on `ComponentFactory`. `ComponentFactory` is now responsible for retaining all live instances of subgraphs (aka subcomponents) that it creates via the `build(:)` function. Subgraphs will be deinitialized alongside `ComponentFactory`.