Add swiftc-ast wrapper script.
This script is intended to be used as a proxy for compiling swift files and emitting their typed AST via `-dump-ast` mode.

Using the script in Xcode involves setting the `SWIFT_EXEC` build setting to the path of this script. Its output file will be based on the `DAST_FILE` preprocessor macro.