[cleansec] Sanitize -dump-ast spaced newlines.
Some `-dump-ast` outputs have an irregularity where particular nodes have incorrect indentation, thus causing them be grouped under the wrong parent. This irregularity materializes as a newline with `n` whitespace characters that should be prepended to the following line.

This change modifies `sanitize(lines:)` to account for this possibility. Introduces a helper `flatMapScanPrevious` which allows for scanning the previous element within `flatMap`.