Add suggested modules to "Missing Provider" errors.
* If one has a missing provider and has compiled a `Cleanse.Module` that `cleansec` parses from the search paths to include that particular provider, we will include a note below the error message with a list of suggested modules they might wish to include.

Depending on the user's search path for parsed modules, it's possible that the suggested module could be stale and not have been _actually_ compiled/updated. This will likely occur if the output path for all cleansec modules is the same. Considering this is merely a suggestion to help guide the user towards finding their module, this edge case doesn't pose too much of an issue.

* Added `ResolutionErrorFormatTests` to add coverage for the emitted error description that goes to `stderr` that Xcode parses.